fix: Fix typescript types
