+++ server.d.ts
@@ -1,0 +1,6 @@
+// TypeScript Version: 2.9
+
+export function collect(
+  html: string,
+  css: string
+): { critical: string; other: string };
