+++ tslint.json
@@ -1,0 +1,7 @@
+{
+  "extends": "dtslint/dtslint.json",
+  "rules": {
+    "interface-over-type-literal": false,
+    "strict-export-declare-modifiers": false
+  }
+}
