--- types/tslint.json
@@ -1,6 +1,0 @@
-{
-  "extends": "dtslint/dtslint.json",
-  "rules": {
-    "interface-over-type-literal": false
-  }
-}
